Renamed and moved files in s3_datasets_shares
