+++ tests/modules/s3_datasets_shares/tasks/test_s3_access_point_share_manager.py
@@ -10,7 +10,7 @@ from dataall.core.environment.db.environment_models import Environment, Environm
 from dataall.core.organizations.db.organization_models import Organization
 from dataall.modules.s3_datasets_shares.aws.s3_client import S3ControlClient
 from dataall.modules.shares_base.db.share_object_models import ShareObject, ShareObjectItem
-from dataall.modules.s3_datasets_shares.services.managed_share_policy_service import SharePolicyService
+from dataall.modules.s3_datasets_shares.services.s3_share_managed_policy_service import S3SharePolicyService
 from dataall.modules.s3_datasets_shares.services.share_managers import S3AccessPointShareManager
 from dataall.modules.s3_datasets.db.dataset_models import DatasetStorageLocation, S3Dataset
 from dataall.modules.shares_base.services.sharing_service import ShareData
@@ -334,11 +334,11 @@ def test_grant_target_role_access_policy_test_empty_policy(
         return_value=None,
     )
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -373,7 +373,7 @@ def test_grant_target_role_access_policy_test_empty_policy(
     # When
     share_manager.grant_target_role_access_policy()
 
-    expected_policy_name = SharePolicyService(
+    expected_policy_name = S3SharePolicyService(
         environmentUri=target_environment.environmentUri,
         role_name=share1.principalIAMRoleName,
         account=target_environment.AwsAccountId,
@@ -400,11 +400,11 @@ def test_grant_target_role_access_policy_existing_policy_bucket_not_included(
 
     mocker.patch('dataall.base.aws.iam.IAM.get_managed_policy_default_version', return_value=('v1', iam_policy))
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -424,10 +424,10 @@ def test_grant_target_role_access_policy_existing_policy_bucket_not_included(
 
     # Iam function is called with str from object so we transform back to object
     policy_object = iam_policy
-    s3_index = SharePolicyService._get_statement_by_sid(
+    s3_index = S3SharePolicyService._get_statement_by_sid(
         policy=policy_object, sid=f'{IAM_S3_ACCESS_POINTS_STATEMENT_SID}S3'
     )
-    kms_index = SharePolicyService._get_statement_by_sid(
+    kms_index = S3SharePolicyService._get_statement_by_sid(
         policy=policy_object, sid=f'{IAM_S3_ACCESS_POINTS_STATEMENT_SID}KMS'
     )
 
@@ -440,7 +440,7 @@ def test_grant_target_role_access_policy_existing_policy_bucket_not_included(
     )
 
     # Assert that statements for S3 bucket sharing are unaffected
-    s3_index = SharePolicyService._get_statement_by_sid(policy=policy_object, sid=f'{IAM_S3_BUCKETS_STATEMENT_SID}S3')
+    s3_index = S3SharePolicyService._get_statement_by_sid(policy=policy_object, sid=f'{IAM_S3_BUCKETS_STATEMENT_SID}S3')
 
 
 def test_grant_target_role_access_policy_existing_policy_bucket_included(mocker, share_manager):
@@ -452,12 +452,12 @@ def test_grant_target_role_access_policy_existing_policy_bucket_included(mocker,
     mocker.patch('dataall.base.aws.iam.IAM.get_managed_policy_default_version', return_value=('v1', iam_policy))
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -903,12 +903,12 @@ def test_delete_target_role_access_policy_no_remaining_statement(
         'dataall.base.aws.iam.IAM.get_managed_policy_default_version', return_value=('v1', existing_target_role_policy)
     )
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -928,7 +928,7 @@ def test_delete_target_role_access_policy_no_remaining_statement(
     # When we revoke IAM access to the target IAM role
     share_manager.revoke_target_role_access_policy()
 
-    expected_policy_name = SharePolicyService(
+    expected_policy_name = S3SharePolicyService(
         environmentUri=target_environment.environmentUri,
         role_name=share1.principalIAMRoleName,
         account=target_environment.AwsAccountId,
@@ -1006,12 +1006,12 @@ def test_delete_target_role_access_policy_with_remaining_statement(
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -1032,7 +1032,7 @@ def test_delete_target_role_access_policy_with_remaining_statement(
     share_manager.revoke_target_role_access_policy()
 
     # Then
-    expected_policy_name = SharePolicyService(
+    expected_policy_name = S3SharePolicyService(
         environmentUri=target_environment.environmentUri,
         role_name=share1.principalIAMRoleName,
         account=target_environment.AwsAccountId,
@@ -1199,12 +1199,12 @@ def test_check_bucket_policy_missing_sid(mocker, base_bucket_policy, share_manag
 def test_check_target_role_access_policy(mocker, share_manager):
     # Given
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -1235,12 +1235,12 @@ def test_check_target_role_access_policy(mocker, share_manager):
 def test_check_target_role_access_policy_existing_policy_bucket_and_key_not_included(mocker, share_manager):
     # Given
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -1272,12 +1272,12 @@ def test_check_target_role_access_policy_existing_policy_bucket_and_key_not_incl
 def test_check_target_role_access_policy_test_no_policy(mocker, share_manager):
     # Given
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=False,
     )
 
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=True,
     )
 
@@ -1293,12 +1293,12 @@ def test_check_target_role_access_policy_test_no_policy(mocker, share_manager):
 def test_check_target_role_access_policy_test_policy_not_attached(mocker, share_manager):
     # Given
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_exists',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_exists',
         return_value=True,
     )
     # Policy is not attached
     mocker.patch(
-        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.SharePolicyService.check_if_policy_attached',
+        'dataall.modules.s3_datasets_shares.services.managed_share_policy_service.S3SharePolicyService.check_if_policy_attached',
         return_value=False,
     )
